RAG Study/rag_chatbot_v4/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
python-dotenv
sentence-transformers
faiss-cpu
fastapi 
uvicorn 
python-multipart
